refactor(tcc): fold the two chroot passes into their driver packages
The last survivors of the numbered-pass pipeline were pass3/ (compile chroot)
and decompile/pass1/ (chroot unwrap) — each a single generic transform behind a
trivial Pipe(). Move the compile chroot into package interp (chrootEnvelope) and
the unwrap into package decompile (unwrapEnvelope), delete both folders. No
passX packaging remains; compiled output is byte-identical.
