+++ pkg/tcc/interp/chroot_test.go
@@ -1,4 +1,4 @@
-package pass3
+package interp
 
 import (
 	"context"
@@ -10,30 +10,22 @@ import (
 )
 
 func TestChroot_WrapsInObject(t *testing.T) {
-	// Setup: Create a processed config object
 	obj := object.New[object.Refrence]()
 	obj.Set("id", "project-id-123")
 	funcsObj, _ := obj.CreatePath("functions")
 	funcSel := funcsObj.Child("func-id-456")
 	funcSel.Set("name", "myFunction")
 
-	// Execute: Run chroot transformer
-	transformer := Chroot()
 	ctx := transform.NewContext[object.Refrence](context.Background(), obj)
-	result, err := transformer.Process(ctx, obj)
-
-	// Verify: Object wrapped in "object" container
+	result, err := (&chroot{}).Process(ctx, obj)
 	assert.NilError(t, err)
 
-	// Verify "object" child exists
 	objectChild, err := result.Child("object").Object()
 	assert.NilError(t, err)
 
-	// Verify original content is in the object child
 	id := objectChild.Get("id")
 	assert.Equal(t, id.(string), "project-id-123")
 
-	// Verify functions are in the object child
 	_, err = objectChild.Child("functions").Object()
 	assert.NilError(t, err)
 }
@@ -41,14 +33,11 @@ func TestChroot_WrapsInObject(t *testing.T) {
 func TestChroot_EmptyObject(t *testing.T) {
 	obj := object.New[object.Refrence]()
 
-	transformer := Chroot()
 	ctx := transform.NewContext[object.Refrence](context.Background(), obj)
-	result, err := transformer.Process(ctx, obj)
-
+	result, err := (&chroot{}).Process(ctx, obj)
 	assert.NilError(t, err)
 	assert.Assert(t, result != nil)
 
-	// Verify "object" child exists even for empty object
 	_, err = result.Child("object").Object()
 	assert.NilError(t, err)
 }
@@ -57,7 +46,6 @@ func TestChroot_WithNestedResources(t *testing.T) {
 	obj := object.New[object.Refrence]()
 	obj.Set("id", "project-id")
 
-	// Add multiple resource types
 	funcsObj, _ := obj.CreatePath("functions")
 	funcsObj.Child("func1").Set("id", "id1")
 
@@ -67,16 +55,13 @@ func TestChroot_WithNestedResources(t *testing.T) {
 	websitesObj, _ := obj.CreatePath("websites")
 	websitesObj.Child("website1").Set("id", "id3")
 
-	transformer := Chroot()
 	ctx := transform.NewContext[object.Refrence](context.Background(), obj)
-	result, err := transformer.Process(ctx, obj)
-
+	result, err := (&chroot{}).Process(ctx, obj)
 	assert.NilError(t, err)
 
 	objectChild, err := result.Child("object").Object()
 	assert.NilError(t, err)
 
-	// Verify all resources are in the object child
 	_, err = objectChild.Child("functions").Object()
 	assert.NilError(t, err)
 
@@ -84,5 +69,18 @@ func TestChroot_WithNestedResources(t *testing.T) {
 	assert.NilError(t, err)
 
 	_, err = objectChild.Child("websites").Object()
+	assert.NilError(t, err)
+}
+
+func TestChrootEnvelope_WrapsProject(t *testing.T) {
+	obj := object.New[object.Refrence]()
+	obj.Set("id", "project-id-123")
+
+	ctx := transform.NewContext[object.Refrence](context.Background(), obj)
+	result, err := chrootEnvelope().Process(ctx, obj)
+	assert.NilError(t, err)
+	assert.Assert(t, result != nil)
+
+	_, err = result.Child("object").Object()
 	assert.NilError(t, err)
 }
