--- pkg/tcc/interp/decompile/pass1/pipe.go
@@ -1,13 +1,0 @@
-package pass1
-
-import (
-	"github.com/taubyte/tau/pkg/tcc/interp/utils"
-	"github.com/taubyte/tau/pkg/tcc/object"
-	"github.com/taubyte/tau/pkg/tcc/transform"
-)
-
-func Pipe() []transform.Transformer[object.Refrence] {
-	return []transform.Transformer[object.Refrence]{
-		utils.Global("", Chroot()),
-	}
-}
